pyproject.toml
[project]
name = "dc-activity-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "click>=8.1.8",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "twilio>=9.4.4",
]
